refactor(ConfirmOrderPage):S-dot 컨벤션 적용
